pyproject.toml
[project]
name = "breaking-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "mysql-connector>=2.2.9",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "pip>=25.3",
    "playwright>=1.56.0",
    "requests>=2.32.5",
]
